feat: wire in confirmModal into Settings page, and delete SettingsContainer
